Safari-609.1.11.1/LayoutTests/svg/css/case-sensitive-tags-expected.txt

PASS window.getComputedStyle(document.getElementById("div1")).color is "rgb(0, 128, 0)"
PASS window.getComputedStyle(document.getElementById("div2")).color is "rgb(0, 128, 0)"
PASS window.getComputedStyle(document.getElementById("div3")).color is "rgb(0, 128, 0)"
PASS window.getComputedStyle(document.getElementById("fo1")).color is "rgb(0, 128, 0)"
PASS window.getComputedStyle(document.getElementById("fo2")).color is "rgb(0, 128, 0)"
PASS window.getComputedStyle(document.getElementById("fo3")).color is "rgb(0, 128, 0)"
PASS window.getComputedStyle(document.getElementById("fo4")).color is "rgb(0, 128, 0)"
PASS window.getComputedStyle(document.getElementById("circle1")).fill is "rgb(0, 128, 0)"
PASS window.getComputedStyle(document.getElementById("circle2")).fill is "rgb(0, 128, 0)"
PASS window.getComputedStyle(document.getElementById("circle3")).fill is "rgb(0, 128, 0)"
PASS window.getComputedStyle(document.getElementById("stop1")).stopColor is "rgb(0, 128, 0)"
PASS window.getComputedStyle(document.getElementById("stop2")).stopColor is "rgb(0, 128, 0)"
PASS window.getComputedStyle(document.getElementById("stop3")).stopColor is "rgb(0, 128, 0)"
PASS window.getComputedStyle(document.getElementById("stop4")).stopColor is "rgb(0, 128, 0)"
PASS window.getComputedStyle(document.getElementById("stop5")).stopColor is "rgb(0, 128, 0)"
PASS window.getComputedStyle(document.getElementById("stop6")).stopColor is "rgb(0, 128, 0)"
PASS window.getComputedStyle(document.getElementById("stop7")).stopColor is "rgb(0, 128, 0)"
PASS window.getComputedStyle(document.getElementById("stop8")).stopColor is "rgb(0, 128, 0)"
PASS successfullyParsed is true

TEST COMPLETE
